Updated some files

+++ src/java/com/topic/deleteTopic.java
@@ -16,6 +16,7 @@
 package com.topic;
 
 import com.connect.DatabaseConnection;
+import com.string.validateInput;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -43,42 +44,61 @@ public class deleteTopic extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        int topicId = Integer.parseInt(request.getParameter("topicId"));
 
-        try (Connection con = DatabaseConnection.makeConnection()) {
-            /**
-             * ***************
-             * deleting the question tag form question_tag_table
-             * *************************
-             */
+        validateInput input = new validateInput();
 
-            try (PreparedStatement ps = con.prepareStatement("delete from question_topic_tag where tag_id = ?")) {
-                ps.setInt(1, topicId);
-                ps.execute();
+        String message = null;
+        int pageNo = 0;
+        try {
+            int topicId = input.getInputInt(request.getParameter("topicId"));
+            pageNo = input.getInputInt(request.getParameter("pageNo"));
+
+            if (topicId != 0) {
+                DatabaseConnection connection = new DatabaseConnection();
+                try (Connection con = DatabaseConnection.makeConnection()) {
+                    /**
+                     * ***************
+                     * deleting the question tag form question_tag_table
+                     * *************************
+                     */
+
+                    try (PreparedStatement ps = con.prepareStatement("delete from question_topic_tag where tag_id = ?")) {
+                        ps.setInt(1, topicId);
+                        ps.execute();
+                    }
+
+                    /**
+                     * *****
+                     * Deleting the followers detail *********************
+                     */
+                    try (PreparedStatement ps = con.prepareStatement("delete from topic_followers_detail where topic_id = ?")) {
+                        ps.setInt(1, topicId);
+                        ps.execute();
+                    }
+                    /**
+                     * ************
+                     * delteting topic from topic table ************
+                     */
+
+                    try (PreparedStatement ps = con.prepareStatement("delete from topic where unique_id = ?")) {
+                        ps.setInt(1, topicId);
+                        ps.execute();
+                    }
+
+                    message = "Topic has been deleted";
+                } catch (SQLException msg) {
+                    Logger.getLogger(deleteTopic.class.getName()).log(Level.SEVERE, null, msg);
+                }
+            } else {
+                message = "Topic Id is not valid or Zero";
             }
 
-            /**
-             * *****
-             * Deleting the followers detail *********************
-             */
-            try (PreparedStatement ps = con.prepareStatement("delete from topic_followers_detail where topic_id = ?")) {
-                ps.setInt(1, topicId);
-                ps.execute();
-            }
-            /**
-             * ************
-             * delteting topic from topic table ************
-             */
-
-            try (PreparedStatement ps = con.prepareStatement("delete from topic where unique_id = ?")) {
-                ps.setInt(1, topicId);
-                ps.execute();
-            }
-            
+        } catch (SQLException | ClassNotFoundException ex) {
+            Logger.getLogger(deleteTopic.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            request.setAttribute("pageNo", pageNo);
+            request.setAttribute("message", message);
             request.getRequestDispatcher("Admin/topic.jsp").forward(request, response);
-            
-        } catch (SQLException msg) {
-            Logger.getLogger(deleteTopic.class.getName()).log(Level.SEVERE, null, msg);
         }
     }
 }
